Changed the way the game renders to be less of a hack
The game used to render on a double buffer from Java's AWT library, now
the code has switched to only using swing, which is already double
buffered, and the frame class now just overrides the paintComponent()
method, and the render loop calls frame.repaint() (from inside the
camrea class). This change allows for better compatibility with the
Swing library, including removing the need to calculate the size of the
window border to reposition the frame. Now it is automatically
positioned inside the window